Now handles nested commits.
We want to be able to have one connection wrap another connection in a
transaction. In MySQL running BEGIN causes any pending transaction to be
commited and we don't want the wrapping transaction commited. So what we
do is keep track of whether we are inside a transaction already and if
so, we change the transaction commands as follows:
BEGIN -> SAVEPOINT mysql_conc_<client port>
ROLLBACK -> ROLLBACK TO SAVEPOINT mysql_conc_<client port>
COMMIT -> RELEASE SAVEPOINT mysql_conc_<client port>

To the second connection that looks like the same behavior as the normal
transactions, but it doesn't interfere with the wrapping transaction on
the first connection.