Fixed a problem with result set responses.
It turns out that in the row data responses, you can have an 0x00 which
should mean an OK packet, but an OK packet isn't a valid row data
response. Also you could get an 0xfe in there as well, so check and see
if the length of the packet is less than 10 (payload less than 6) and if
it isn't then it can't be an EOF. This is all from the protocol docs and
I was actually getting some packets with 0x00.