Fixed socket write buffer full problem.
Sometimes when we write to the socket, it doesn't buffer everything we
give it, and we need to wait for the buffer to clear out to write again.
I had the code all setup to do this, but when I overwrode the
want_to_write methods in the MySQL and Client connections, I left out
the part that checks if there's anything remaining in the write buffer.
That was causing things to stall out because it wasn't getting added to
the write queue.

Also limiting the max write to 32768 because that seems to be the
behaviour of the PHP client.