Handle disconnection during auth better.
If the client disconnected during auth, the concentrator would be stuck
in a state waiting for the auth reposonse from the client and wouldn't send
the handshake init message to any new clients that connected in. Also,
it was just throwing an exception if authentication failed. Now we will just
forward on the error message to the client and treat it as if auth failed
and then the next client can try and connect.